chore(validate): rename file `validations.js` to `validate.js`
